# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "ScholarMe API Specification",
    "description": "Authoritative OpenAPI 3.0 specification for ScholarMe core API route groups."
  },
  "servers": [
    {
      "url": "https://scholarme.vercel.app"
    }
  ],
...
  "paths": {
...
    "/api/auth/register-card": {
      "post": {
        "description": "Register or link an HMAC-signed QR ID card to a user profile (Admin/Officer only)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "cardId": {
                    "type": "string",
                    "example": "CARD-12345"
                  },
                  "pin": {
                    "type": "string",
                    "minLength": 4,
                    "maxLength": 4
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Card successfully linked to profile"
          },
          "403": {
            "description": "Forbidden: Officer or Admin privileges required"
          }
        }
      }
    },
...
  }
}